Fix: fenced code prefix regression
Flexmark automatically adds prefix language- if code
fenced between three ticks had a language prefix.

This is not backwards-compatible with the library Cryogen
used to use. If someone enabled Klipse on fenced code, with
a selector based on the class, Klipse would no longer work
after the markdown parser upgrade.